+++ CHANGELOG.md
@@ -9,10 +9,17 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Introduce `std` (enabled by default) and `unstable` features ([#58]).  They
+  enable corresponding serde’s features and if either is enabled, `Error`
+  implements `std::error::Error` trait.  By itself, `serde-json-wasm` is now
+  `no_std`; it’s up to serde’s features whether the entire build is.  **Please
+  not:** this potentially breaks `default-features = false` builds.
+
 - Serialize / deserialize `u128`/`i128` types as numbers instead of strings
   ([#59]).<br/> **Please note:** this breaks deserialization of `u128`/`i128`
   serialized with older versions of `serde-json-wasm`.
 
+[#58]: https://github.com/CosmWasm/serde-json-wasm/pull/58
 [#59]: https://github.com/CosmWasm/serde-json-wasm/pull/59
 
 ## [0.5.1] - 2023-04-11